Fix uint48 cast for getProposalHash parameter
The getProposalHash function expects a uint48 parameter, but
_lastProposalId is uint256. Added explicit cast to fix compilation.

+++ packages/protocol/contracts/layer1/core/impl/Inbox.sol
@@ -520,7 +520,7 @@ contract Inbox is IInbox, IForcedInclusionStore, EssentialContract {
         view
     {
         bytes32 hashToProve = LibHashOptimized.hashProveInput(
-            getProposalHash(_lastProposalId), _input
+            getProposalHash(uint48(_lastProposalId)), _input
         );
         _proofVerifier.verifyProof(_proposalAge, hashToProve, _proof);
     }
